Use crypto-js instead of crypto for DES encryption
MongoDB Realm does not support native DES encryption using the node.js crypto library. I suspect this has to do with security reasons and disabling DES in the native OpenSSL library on MongoDB Realm machines.

To circumvent this problem we use crypto-js for DES encryption. crypto-js implements DES in pure javascript, so that MongoDB Realm should not have a problem with this.